added sign up styles to a seperate file
